Move key manager error to module, fix linting
