+++ bindings/uniffi/src/dids/resolution/mod.rs
