+++ crates/web5/src/apid/dids/resolution/resolution_metadata.rs
@@ -30,4 +30,3 @@ pub enum ResolutionMetadataError {
     #[serde(rename = "internalError")]
     InternalError,
 }
-
